% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{conversion_table}
\alias{conversion_table}
\alias{convert}
\title{conversion_table}
\usage{
convert(
  values,
  from = "isco_88",
  to = "egp",
  selfEmployed = NULL,
  unemployed = NULL
)
}
\arguments{
\item{values}{vector of values for conversion}

\item{from}{origin code name}

\item{to}{destination code name}

\item{selfEmployed}{vector of origin codes for self-employed}

\item{unemployed}{vector of origin codes for unemployed}
}
\value{
Data table of oring and destination codes
}
\description{
conversion_table
}
\examples{
convert(values = c("9321","9322","9330","1210"), from="isco_88", to = "isei")
}
